Decouple main and test
